♻️  Refactor: StoreLocation 관련 파일 구조 분리 및 디렉토리 정리
- StoreLocationView, ViewModel을 Components/StoreLocation으로 이동
- Marker, Place, MapViewAdapter 등을 Components/Map으로 이동
- MapRepresentable 프로토콜은 Common/Protocol/Map 폴더로 이동
- 모듈 간 책임 분리를 위해 구조 개선
